Merge in the mca_base_select changes per RFC:  http://www.open-mpi.org/community/lists/devel/2008/04/3779.php
{{{
svn merge -r 18276:18380 https://svn.open-mpi.org/svn/ompi/tmp-public/jjh-mca-play .
}}}

Any components not in the trunk, but in one of the effected frameworks *must* be
updated. Contact the list, look at the RFC, or look at the diff for how to do this.

Sorry for the early commit of this, but I wanted to get it in today (per RFC) and
didn't know if I would have a chance later today.



git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@18381 63e3feb5-37d5-0310-a306-e8a459e722fe