Abstract iWARP subnet ID functions (sans build break)
The iWARP subnet ID determination should not be in the RDMACM cpc, as
it was in the preversion, as this violates the cpc abstract that is
present throughout the code.  Also, this patch uses the opal_list_t
data struct instead of using its own linked lists.

This attempt includes *iwarp.c and *iwarp.h


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@18414 63e3feb5-37d5-0310-a306-e8a459e722fe