Make sure to check "opal_list_get_last" instead of "opal_list_get_end".
The former will return a valid item in the list, the latter will return an invalid item that marks the end of the list.

It was happending that when oversubscribing by way of an appfile we would cause a segv because we tried to interpret the invalid item returned by "opal_list_get_end" instead of a valid item. We would then try to write to unallocated memory.

This commit fixes #1279



git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@18529 63e3feb5-37d5-0310-a306-e8a459e722fe