Fixes #1305: check to see if $sysfsdir/class/infiniband exists and is non-empty.  If not, then exit the openib btl silently.  This addresses the case where libibverbs is installed (which is getting more common) and therefore the openib BTL was built/installed, but the kernel drivers are not loaded (assumedly because there is no RDMA hardware present).  In this case, "mpirun a.out" will not issue a warning.
There appears to be no good way to definitely tell if there are no
RDMA hardware devices present.  For example, if libibverbs/the openib
BTL is installed, there are no RDMA devices present, but the RDMA
hardware kernel drivers ''are'' loaded, OMPI will warn that it was
unable to find suitable devices.  This warning is easily eliminated by
unloading the kernel drivers.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@18530 63e3feb5-37d5-0310-a306-e8a459e722fe