Properly ensure that cmd line MCA params override environmental MCA params for procs local to mpirun.
Actually, the problem was that we were simply -adding- any enviro MCA params to whatever had been found on the cmd line. Thus, duplicate MCA param directives were winding up duplicated in the environment. Some shells took the first one in the environ array - others took the last! So we could get completely different behavior based on the whims of the shell.

This commit fixes #1373


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@18836 63e3feb5-37d5-0310-a306-e8a459e722fe