Fixes #1135.
Short version: when the HNP launches VPID 0 on the same node as
itself, the STDIN IOF endpoint will have both a pub and a sub on it.
We need to ensure to only forward incoming messages ''once'' (not
twice, as was happening).  A lengthy comment in the code explains in
more detail.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@18873 63e3feb5-37d5-0310-a306-e8a459e722fe