Per July technical meeting:
Revise the scope precedence in the MPI_Publish, Unpublish, and Lookup functions. If a global server was specified and is available, then default to using it for all three functions. If not, then default to using local scope.

If an info_key was provided, then it takes preference. We always follow the user's direction - this change only impacts the scope ordering if the user -doesn't- tell us the order to use.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@19146 63e3feb5-37d5-0310-a306-e8a459e722fe