[Re-]Fix #1463 with a little thing that I like to call "the right way".
Don't modify coords in the top-level API function because coords is an
IN variable.  Instead, as Nysal noted, the real cause of the problem
was a missing ! down in topo_base_cart_rank.c.  Put a comment down in
topo_base_cart_rank.c explaining what's going on so that the code is
not so cryptic.

Refs #1363.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@19487 63e3feb5-37d5-0310-a306-e8a459e722fe