Replace the old coordinated component ('coord') and replace it with a much more refined version ('bkmrk').
The new component fixes a number of problems with the old component. The core algorithm is the same, but by changing the data strucutres a bit we have improved performance and memory utilization.

There are still a couple corner cases that still need some work. However, I did not want to delay bringing this into the trunk (and v1.3 branch) for too much longer.




git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@19537 63e3feb5-37d5-0310-a306-e8a459e722fe