Only set ompi_mpi_leave_pinned (a bool) to true if the MCA param value is >= 1.  The default value of the MCA param is now -1, which means "let someone else turn it on if they want to."  So we should default to ''off'' (false), and let the openib BTL (etc.) turn it on if it can/wants to.
Failure to do this will default _pipeline to true because
-1(int)==true(bool).  This causes a problem if the user tries to set
mpi_leave_pinned_pipeline to 1: they'll get a warning that you can't
set both _pinned and _pinned_pipeline to 1.  This happens because
_pinned will get the bool-ified value of of the MCA parameter (-1),
and then the user sets the value of _pinned_pipeline to 1/true.
Hence, both of them are set to true.  Bzzt!


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@19942 63e3feb5-37d5-0310-a306-e8a459e722fe