A fix for the intel "MPI_Send_init_ator_c" test.
It highlighted a bug in the bookmark component where for persistent sends we were not copying the context, but just moving it. This caused us to lose track of the message if it is started/completed multiple times.

This will need to be brought over to the v1.3 branch, but it should soak overnight to get a round of testing first.



git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@19962 63e3feb5-37d5-0310-a306-e8a459e722fe