This disables sendi, since it may do 0-byte requests and it still has another bug.  This also causes 0-byte requests to be treated as a buffer error, causing the base request to be requeued.  On Cray XT, it may be temporarily impossible to make allocations for buffer requests, as the default stack size is small (8 MB) and there is no true swap device. Even with the stack size increased, there will be cases in which this condition recurs.
One possibility is to make the buffer allocations off of the heap; but,
this does not change the fact that eventually an out-of-memory condition
will occur and we need to support multiple receives in transit, a
condition for which the available buffer space may change.  On the other
hand, if we switch to allocating the buffer space from the heap, we will
need to return an error when the allocation fails and there are no other
buffers in transit.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@19981 63e3feb5-37d5-0310-a306-e8a459e722fe