Modify make_dist_tarball to conditionally allow it when ''higher'' than expected versions of tools are found.  make_dist_tarball will ''not'' accept lower or higher versions by default (since the goal is exact reproducability, after all), but you can specify the --highok command line switch to allow it.
I'm also removing make_tarball in this commit and will replace it with
a sym link to make_dist_tarball in the next commit.  The script will
default to allowing higher versions of the tools when invoked with an
argv[0] of "make tarball" (i.e., --highok is not necessary, but is
harmless).


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@21530 63e3feb5-37d5-0310-a306-e8a459e722fe