Correct a couple of bugs in the rank_file mapper that were incorrectly assigning vpids.
Add a capability to parse the rankfile to extract node information in place of requiring both hostfile and rankfile for non-RM managed environments. The rankfile is -only- parsed for this IF the hostfile and -host options are not given. Otherwise, those are used to establish allocation info as we did before this commit.



git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@21815 63e3feb5-37d5-0310-a306-e8a459e722fe