Closes #2158.
Special-case the before MPI_INIT / after MPI_FINALIZE error messages
so that they can be a bit more clear than the general "an error
occurred" messages that are displayed in the middle of MPI jobs.

This is not really a "bug fix", but it is helpful for usability.  I
leave it up to the v1.4 RM's to decide if they want it for the 1.4
series or not.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@22382 63e3feb5-37d5-0310-a306-e8a459e722fe