Add a patch to libltdl so that the preopen loader doesn't mask a real error if the file-based dloader (e.g., dlopen) fails to load a DSO for a complex reason (e.g., unresolved symbol).
This has been widely reported upstream to the libltdl maintainers - a
general solution is difficult.  This is very definitely an
OMPI-specific solution.  Since our embedded libltdl is hidden behind
visibility flags, that's ok.

Note that this is a change to autogen.sh, but this commit does not
force re-running autogen.sh.  You'll just get the new functionality
the next time you re-run autogen.sh.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@22806 63e3feb5-37d5-0310-a306-e8a459e722fe